fix: 🐛 rename font repos to resolve repo not found error
Prior to this change, the fedora build complained that the font.repo
was unable to be found. I had mistankanly thought that the ostree
configuration used the name of file of the to set the repo path. However
it uses the name to configure this.

After this change, the font.repo was renamed to peterwu/iosevka to reflect
the same name set in the repo file